@@ -1,10 +1,17 @@
 cmake_minimum_required (VERSION 3.10)
 
-set (CMAKE_CXX_STANDARD 11)
+set (CMAKE_CXX_STANDARD 17)
 
 project (net)
 
 aux_source_directory(./src  DIR_SRCS)
+
+if(USE_SSL)
+  add_definitions("-D__ENABLE_SSL")
+endif()
+
+add_subdirectory(test)
+add_subdirectory(examples)
 
 if(${CMAKE_SYSTEM_NAME} MATCHES "Linux")
   list(FILTER DIR_SRCS EXCLUDE REGEX ".net_kqueue.*")
@@ -18,7 +25,6 @@ add_dependencies(net
   protobuf
 )
 
-message(STATUS "protobuf include dir: ${PROTOBUF_INCLUDE_DIR}")
 target_include_directories(net
   PUBLIC ${PROJECT_SOURCE_DIR}/..
   PUBLIC ${INSTALL_INCLUDEDIR})
